Add rewind() support to the MediaSourceStream.
Rewinding allows a stream to be backtracked without touching the underlying source.
Unseekable sources are therefore given a modest capability for backwards seeking.
This functionality is provided in lieu of "peek" functions, and is more powerful.

MediaSourceStream also gains an exponentially growing read-ahead buffer to reduce
the overhead of seeks.