Load tags into an array in FormatReader.
Previously, the idea was that an event hook would be fired when a tag
was encountered and then the user could decide what to do with it:
skip it, read it all, or read some of it. However, this won't work
when in the future things could be running on different threads.
Limits will be used to prevent denial of service attacks in this
case.