metadata: Add crate and move over relevant modules from core.
Just like codec and format decoders are not in the core sonata crate,
tag readers and writers are also not to be implemented in the core
crate. All tag readers and writers will be in the metadata crate
going forward.

The metadata crate will contains all metadata readers and writers.