flac: Flush fragment queue on IO error before returning the error.
The FLAC frame parser reconstructs FLAC frames from FLAC frame fragments.
A fragment is a buffer that starts with a legitimate looking FLAC frame
header and ends before the next frame header. Since it's possible that
a valid FLAC frame contains the same byte pattern as a frame header within
its audio data, multiple fragments may need to be combined together to
reconstruct a FLAC frame.

Fragments are read from the MediaSourceStream into a queue. They are then
scored and combined. However, there are two edge-cases:

1) When the fragment queue only contains 1 fragment, it's difficult to be
   certain that it, itself, is a full FLAC frame.

2) The MediaSourceStream will return an end-of-stream error before the
   fragment queue is emptied.

To solve these edge cases, keep the fragment queue filled with atleast
2 fragments at any given time. If the MediaSourceStream returns an error,
instead of returning it immediately, store it and use it as a flag to
fully empty the fragment queue. Once the fragment queue is emptied, return
the error.