mp4: Improve abstraction of fragmented and non-fragmented streams.
The initial attempt of abstracting fragmented and non-fragmented streams
included pre-processing and converting atoms into a unified internal
representation. This representation was poor for both types of streams in
certain scenarios.

Rewrite the abstraction such that the original atoms, moov and moof (for
non-fragmented and fragmented streams, respectively) can be used as-is
by wrapping them in thin structs implementing a new trait, StreamSegment.

Further work is required to improve the performance of fragmented streams
in certain worst-case scenarios.