mp3: Sample accurate slow-seeking.
Implement sample accurate slow-seeking for mp3 elementary streams.
Slow-seeking works by parsing frame-by-frame until the frame containing
the desired sample is found. If the desired timestamp is in the future,
then frame parsing begins at the current position. If the desired
timestamp is in the past, then frame parsing begins at the start of
the stream.

Three future improvements:

1) Build a seek index on load and/or while the stream is playing.

2) Take into account the bit resevoir to ensure that it will contain
   the correct data when the decoder gets to the frame containing the
   sample specified by the seek timestamp.

3) Implement fast-seeking where the position of the correct frame is
   guessed based on the average bitrate.