core: Optimize MSS::ignore_bytes for large counts.
Ignore bytes consumes count bytes from the source by reading and
then discarding the read bytes. For large counts this is very slow
compared to a seek. Optimize this operation on seekable sources by
first seeking forward majority of the way, then by using the previous
read-and-discard method to reach the desired count. Note that since
ignoring bytes is a rewindable operation and that a seek invalidates
the rewind/ring buffer, the amount of bytes that can be skipped by
the seek is count - the ring-buffer length.