wav: Support ffmpeg streaming wav output.
When outputting to stdout, ffmpeg will set both the riff and data
chunk lengths to (2^32)-1 because it can't seek back and write the
correct lengths. This technically makes no sense because the data chunk
is a child of the riff chunk and there are other chunks within the riff
chunk before the data chunk. Thus, the data chunk cannot be the same
size as the riff chunk. WavReader correctly detects this as an error,
however, this case needs to be supported for check. Therefore, detect
this condition and allow it.