core: Rename formats::Stream to formats::Track.
This change is to prevent confusion between IO streams such as
BitStream, ByteStream, MonitorStream, etc. and the sub-streams (aka
streams) of a container format. Since tracks and streams are
largely interchangeable when discussing muxed media formats this
is a good step to reduce the overloading of the word stream in the
library.