core: BitStreams v2 part 1.
This is part of a series of changes to rewrite of Symphonia's bit
reading primitives.

The motivation for this rewrite was to accomplish the following goals:
  1) Increase performance
  2) Support both MSb and LSb bit reading directions
  3) Enforce a naming convention to better describe the operation of
     each bit reading primitive and make it consistent with the byte
     reading primitives.

First, increasing performance. The existing BitReaderLtr and
BitStreamLtr consumed bytes from a ByteStream one byte at a time. This
could not be worked around because a ByteStream can only be advanced,
thus consuming more than one byte at a time was not an option. This is
not ideal for performance since reading a new byte is compartively
expensive vs. extracting bits.

Therefore, in v2, while a BitStream will still operate the same way, a
BitReader references a [u8] and consumes 8 bytes at a time. Since
Packets are fancy [u8] buffers, this should be the primary bit reading
primitive.

For Decoders that are heavily IO bound (e.g., FLAC), this has been
observed to increase the performance of the Decoder by 20-40%.

Second, while the current set of Decoders and Formats prefer reading
bits in MSb-first order, Vorbis will need the ability to read bits
in LSb-first order. Therefore, both directions must be supported. The
v2 rewrite provides equal support for both bit reading directions.

Finally, the nomenclature of the io module is improved and unified
between bit and byte reading primitives. A "Reader" now consumes an
underlying [u8] buffer, while a "Stream" consumes a source that
implements ReadBytes.