core: Optimize AudioPlanes(Mut) for low channel counts.
AudioBuffer supports up-to 32 channels (planes). When getting slices
to the planes of an AudioBuffer via planes() or planes_mut() the
returned AudioPlanes or AudioPlanesMut structures must be capable
of holding up-to 32 slices. This uses 512 bytes of stack space which
is a lot.

Since it's unlikely > 7.1 audio (8 channels) will ever be used,
optimize AudioPlanes and AudioPlanesMut to only store up-to 8 slices
on the stack (128 bytes). If > 8 channels are used, then fallback to
a vector to store the planes. This is still not great and incurs an
allocation in some cases, but better than the previous implementation.