core: BitStreams v2 part 2, variable-length code module.
This is part 2 in a series of changes to rewrite Symphonia's bit
reading primitives.

This change introduces the core::io::vlc module, a generic module
to support variable-length code decoding, and removes the existing
core::io::huffman module. Likewise, the read_huffman function
present on the ReadBits traits are replaced by read_codebook.

This module introduces the concept of a codebook. A codebook is
a lookup table that is generated from a list of code words, the
length of each code word, and the values associated with each code
word.

A codebook is used by a bit source (implementer of ReadBits) to
efficiently decode a variable-length code into a usable value.

Unlike the existing huffman module, a codebook is generated at
runtime and can be tuned based on the code words, lengths, and
values.

Despite the overhead of generating codebooks at runtime, with all
other things being equal, read_codebook is ~7% faster than
read_huffman. It should be noted that for the current set of
decoders, the cost of generating the codebook is one-time and
amortized over the duration of playback for the life of the
application.

A higher uplift can be achieved by tweaking the codebook parameters to
better suit the code words. This has been observed to 2x the
performance gain to ~14%. This change will be in a follow-up commit.

Further work will be considered to implement statically declared
codebooks and a codegen tool.