mp3: Parse Xing/Info tags, and estimate duration.
The Xing/Info tags provide information for VBR files. Using this
information the duration can be determined for VBR files. In cases
where there are no tags, estimate the duration by finding the average
MPEG frame sizes and dividing the size of the stream by it.

The LAME extension to the Xing/Info tags is also parsed and can provide
replay gain information.

Partially addresses #51.