ogg: Rewrite OggReader.
This commit is a rewrite and rearchitecture of the OGG demuxer.

The existing demuxer was too tightly coupled and made it difficult, if not
impossible, to support all the potential use-cases OGG supports.

This rewrite better separates the reading of pages from the OGG physical
stream from the parsing of those pages within the context of a logical stream.
To that end, packet mapping is now completely encapsulated within the logical
stream itself.

Notable additions that this new demuxer now supports include:

* Seeking, including within the active physical stream of a chained OGG media
  stream.
* Reading timestamps and durations for all supported codecs.
* Determining the start time (initial timestamp), start delay, and end delay
  of streams. Thus providing the necessary information for gapless playback.
* Determining the duration of streams, within chained or unchained OGG media
  streams.