core: Allow access to `Decoder`'s last frame.
Adds `Decoder::last_decoded` to expose read access to internal buffers after a call to `decode`. This is useful when dealing with multiple tracks/files to be mixed into a single output buffer, particularly if this output buffer is smaller its input frames (e.g., 20ms vs. 26ms MP3 frames). `AudioBufferRef`s can't be safely stored alongside their parent decoders, so this ensures that any leftover data doesn't need to be preemptively copied out.

Currently, this exposes `buf` in all decoder implementations if the frame count is nonzero, and expects the user not to call `last_decoded` if the previous frame caused a decoding error.