check: Add option to ignore decode errors.
Since a decode error will create a discontinuity in the output
produced by Symphonia (a packet is effectively being dropped),
it is expected that this will create subsequent false failures if
the reference decoder also does not drop this packet since the
pcm output will no longer be aligned.

For codecs that don't specify how to handle corrupted or malformed
streams, this doesn't necessarily constitute an error.

Addresses #68.