isomp4: Properly seek to mdat atom in try_new.
For seekable media source streams, all the atoms in the stream are
scanned before seeking back to the first mdat atom for playback.

To accomplish this correctly, the stream must be seeked to the
start, and then atom iterator must also be reset AND iterated to
the desired atom.

Fixes #73.