core: Rewrite IMDCT in terms of an FFT for accuracy.
Previously, the MDCT module implemented the IMDCT using a
Fast Cosine Transform (FCT). The FCT was used to compute a
type-2 DCT, which got transformed into a type-4 DCT, which in
turn got converted to the IMDCT. At large values of N, the
numerical accuracy decreased and caused many deviations from
other reference decoders.

The IMDCT can also be computed using the Discrete Fourier
Transform (DFT) which, when implemented with a Fast Fourier
Transform (FFT) algorithm, has much better numerical
stability across a large range of N.

Thus, this commit adds two new sub-modules: complex, and fft. The
complex module implements a minimal Complex number type, and the
fft module implements the FFT. The dct module is removed, and
the mdct module now uses a FFT to compute the transform instead
of a DCT.

The FFT implemented is a radix-2 Cooley-Tukey algorithm which is
unrolled for N <= 32. The performance is comparable to the
previous implementation, but could likely be improved by further
tuning, or switching to a split-radix FFT variant.

This addresses most, if not all, of #71.