mp3: Xing tag may be truncated or CRC may not be present.
The Xing tag may not be complete. Workaround this by reading only
up-to the encoder delay and padding fields first, then trying for
the rest.

Additionally, only LAME unconditionally adds the CRC field to the
tag. Other encoders may only do so if the protection bit is set in
the MPEG header.

Addresses #99.