mp3: Use stricter frame sync on try_new.
Sometimes the reader will sync to random data instead of the first
frame of a stream. This results in the reader providing an invalid
packet to the Decoder and causes the SignalSpec of the decoder to
be wrong.

This change adds an extra layer of validation that checks if the
next 4 bytes after the packet is read is a MPEG header. If there
is no header after the packet, then the stream is re-synced again.

Fixes #96.