core/audio/RawSampleBuffer: allow using external storage
GStreamer uses its own Buffers which are handed to the downstream
nodes in the processing graph. The nodes (elements) can be implemented
in mutiple languages and the Buffers can be allocated elsewhere or be
part of a Buffer pool.

RawSampleBuffer provides a convenient mechanism to export decoded
buffers to an unaligned bytes stream. In order to ensure data
safety, the RawSampleBuffer manages its own sample storage and
provides an accessor to obtain an immutable byte slice on the
resulting data.

When used in frameworks such as GStreamer, the byte slice must be
copied to the target Buffer.

The changes in this commit allow using the RawSampleBuffer with
an external byte slice, thus avoiding the last copy.