mkv: Fix forward seeks with read-only MediaSources
Read-only sources were mistakenly calling `Seek::seek` if an MkvReader's `clusters` had been populated, causing a seek error even if the target timestamp occurs later in the stream.

This PR changes the EBML iterator to consume bytes if the source is not seekable *and* the destination can be forward-skipped to, otherwise returning `ForwardOnly`.