flac: New packet parser.
Implement a new faster, simpler, and more robust packet parser
for FLAC.

Like the existing implementation, the new parser consumes data from
the media source stream until it find a valid frame header. The
data consumed upto that header is saved as a packet fragment.

One or more fragments are merged together to form a packet, if the
CRC of the merged fragments matches the CRC in the footer of the
newest fragment. If no packet can be formed, then more fragments are
read from the media source stream.

Heuristics and hard limits are used to ensure that there is an upper
bound on the total size and number of fragments read.

Unlike the previous implementation, this parser optimizes for the
typical case where 1 fragment = 1 packet. These optimizations yield
a 1-5% uplift in decode performance depending on the file.

This packet parser also unifies stream synchronization while seeking
and playing, further reducing complexity.