flac: Reset reader position after false synchronization.
To synchronize the FLAC frame parser a full frame header must
be read after it encounters a FLAC frame preamble. If a preamble
is encountered, but it was actually just random data and not the
start of a new FLAC frame, then synchronization must restart from
one byte after the start of the fake preamble. If not, any bytes
consumed after the fake preamble used to determine it was not a
real FLAC frame would be skipped. If a FLAC frame actually began
in those bytes, then we would miss the entire frame.

Fixes #261.