flac: Allow different sample rates between frames
FLAC encodes for each frame the sample rate of that frame, it is
therefore to be expected that the sample rate changes between frames and
is valid per FLAC's test vectors. This patch removes the sample rate
check in `strict_frame_header_check` and updates the sample rate in the
SignalSpec of the AudioBuffer after decoding a frame.

Ref: https://github.com/ietf-wg-cellar/flac-test-files/blob/main/uncommon/01%20-%20changing%20samplerate.flac