meta: Rewrite tag representation.
A Tag is now a composition of a mandatory RawTag, and an optional
StandardTag.

The RawTag stores the key-value pair in a data format that matches
as close as possible to the how the tag was written. A StandardTag
evolves the StandardTagKey concept by holding the parsed value of
the tag in the enum. In this way, a user will be guaranteed to get
the value of a StandardTag in a known data type and format.