mkv: Support recursion with ElementIterator.
The ElementIterator could not support recursive elements (elements
that can contain nested elements of the same type) since each
recursive step would take an additional &mut of the inner reader.
This new mutable reference would be considered a new type to the
generic Element::read function consuming it, thereby causing an
overflow to occur because the compiler would need to instantiate
a new implementation of Element::read at each recursive step.

The solution to this problem is to not take &mut at each step. To
accomplish this, ElementIterator must only take &mut of the inner
reader if it owns the inner reader. Typically, this would only be
the case for a top-level iterator. For lower-level/child iterators
using a borrowed inner reader, the reference must simply be passed
along to the next child iterator.

This is implemented by abstracting owned vs. borrowed inner readers
from ElementIterator via. the ElementReader trait. ElementReader is
a trait that only requires DerefMut to be implemented with the
constraint Deref::Target implements ReadBytes.

The ElementReader trait is implemented by
OwnedElementReader<B: ReadBytes> and BorrowedElementReader<B: ReadBytes>
for the owning/borrowing cases. In the owning case, Deref(Mut) return a
reference to the owned inner reader, while in the borrowing case they
return the pre-existing reference.

In this way, no matter the depth, ElementIterator will only ever use
a reference at-most one indirection deep.

Since recursion may run the risk of a runtime stack overflow with
malicious or malformed input, a maximum depth limit of 16 is enforced.
Excluding nested chapter elements, the maximum expected depth is 7
for MKV files. With a depth limit of 16, up-to 10 levels of nested
chapters will be supported.