Fix Default implementation for TimeBase
A TimeBase represents symphonia's conversion factor between time.
It does not make sense for it to have a factor of 0, which is also
checked by its constructor TimeBase::new. Unfortunately the Default
implementation for TimeBase is derived, yielding a value of 0 for both
u32 members (numerator and denominator), which is probably not what we
want. This change switches the default to be 1.

Also fixes a typo in the doc comment for TimeStamp.