Add basic support for reading average bitrates
Bitrate is seemingly missing from all song metadata in this
project. Even songs that contain the bitrate in the header (alac,
mp3, possibly more?) don't make this information available to the
user.

This commit stores the bitrate in CodecParameters as
"average_bitrate". This is because some formats like flac have
a variable bitrate. It's better to provide something than to
promise a constant bitrate and be technically wrong as some
decoders do.

This average bitrate is currently only implemented for mp3. But
possibly other codecs might want to implement it. This change
is fully backwards compatible.