Make Ogg FormatReader work for partial MediaSourceStreams
At the moment, the Ogg FormatReader assumes a complete
MediaSourceStream, which allows to seek to the last page.
This assumption breaks for partial decodes (MediaSourceStream::Read may
return UnexpectedEof multiple times and then continues to fetch data
that is available on the next call). This patch allows decoding in chunks,
similar to libogg behavior.

Concretely this means:
1. When reading from a MediaSourceStream, we sync back to the last healthy
   position when experiencing an UnexpectedEof.
2. If the last page was not present when initializing the format reader,
   we compute the end_delay when reaching the last page.

Co-authored-by: Jakob Koschel <jakobkoschel@google.com>