caf: demux packets based packet type instead of compression
Since we can have compressed formats with fixed packets sizes this means
that we need to separate by packet type not by compressed or not.

This replaces the check with a varible packet size check and enables the
demuxing of compressed fixed packets.

Uncompressed data is chunked into MAX_FRAMES_PER_PACKET while compressed
data packets are read one packet at a time.