Make FLAC FormatReader work for partial MediaSourceStreams
At the moment, the FLAC FormatReader assumes a complete MediaSourceStream,
which allows to seek to the last page. This assumption breaks for partial
decodes (MediaSourceStream::Read may return UnexpectedEof multiple times
and then continues to fetch data that is available on the next call).
This patch allows decoding in chunks.

Concretely this means:

When we experience an Eof while packet decoding, we sync the reader to
the last known working position. This has no side effects when we
actually reached the streams end, because the reader will already be at
the end in that case.