Cleanup Cargo.tomls and licenses
This patch makes it so that the MPl-2.0 license is bundled with
each crate (as is best practice for license files). However, cargo
does NOT currently have a better solution than copying the license
to each crate's folder (see discussion here:
https://github.com/rust-lang/cargo/issues/13328), so this patch
copies the LICENSE file to each crate's folder.

This patch also consolidates a lot of the package metadata into the
root Cargo.toml.